Refactor: Split AuthProvider into smaller files
Split the large AuthProvider.tsx file into smaller, more manageable files for improved code organization and maintainability. This includes extracting utility functions and custom hooks to enhance readability and reusability.
